Fresh newsview

+++ Frontend-Internet-newspaper/src/stores/NewsStore.js
@@ -1,9 +1,16 @@
 import { defineStore } from "pinia"
+import { useAuthStore } from "./AuthStore"
 
 export const useNewsStore = defineStore("news", {
     state: () => ({
-        news: [],
+        news: null,
         likes: [],
+        edit: false,
+        modal: false,
+        news_for_edit: null,
+        role: '',
+        edit_post_id: null,
+        filtr_news: null
     }),
     actions: {
         async getnews() {
@@ -22,15 +29,8 @@ export const useNewsStore = defineStore("news", {
                 }
                 const responseData = await response.json()
 
-                 //////////111111111111////////////
-        
-                sessionStorage.setItem('news',  JSON.stringify(responseData))
-                this.news = JSON.parse(sessionStorage.getItem('news'));
-                 //////////111111111111////////////
-                // console.log(JSON.parse(localStorage.getItem('news')))
-                // responseData.forEach((item) => {
-                //     this.news.push(item)
-                // })
+                this.news = responseData
+
                 console.log("Response data:", responseData)
 
             } catch (error) {
@@ -68,6 +68,7 @@ export const useNewsStore = defineStore("news", {
                     alert("Неправильный вход!")
                     throw new Error("Authentication failed")
                 }
+                this.getnews()
                 // const responseData = await response.json()
                 // console.log(responseData)
             } catch (error) {
@@ -75,6 +76,7 @@ export const useNewsStore = defineStore("news", {
             }
         },
         async deleteComment(id_comment) {
+            const role = sessionStorage.getItem("user_role")
             const jwtToken = sessionStorage.getItem("jwtToken")
             console.log(jwtToken)
 
@@ -85,21 +87,38 @@ export const useNewsStore = defineStore("news", {
                 console.log(jwtToken)
             }
             try {
-                const response = await fetch(
-                    `http://localhost:8085/comment/user/${id_comment}`,
+                if (role == 'ROLE_ADMIN') {
+                   const response = await fetch(
+                    `http://localhost:8085/comment/admin/${id_comment}`,
                     {
                         method: "DELETE",
                         headers: new Headers({
                             Authorization: "Bearer " + jwtToken,
                             "Content-Type": "application/json",
                         }),
-                        // body: JSON.stringify(id_com),
                     }
                 )
                 if (!response.ok) {
                     alert("Удаление не прошло")
                     throw new Error("Authentication failed")
+                } 
+                } else if (role == 'ROLE_USER') {
+                    const response = await fetch(
+                        `http://localhost:8085/comment/user/${id_comment}`,
+                        {
+                            method: "DELETE",
+                            headers: new Headers({
+                                Authorization: "Bearer " + jwtToken,
+                                "Content-Type": "application/json",
+                            }),
+                        }
+                    )
+                    if (!response.ok) {
+                        alert("Удаление не прошло")
+                        throw new Error("Authentication failed")
+                    } 
                 }
+                this.getnews()
                 // const responseData = await response.json()
                 // console.log(responseData)
             } catch (error) {
@@ -107,10 +126,6 @@ export const useNewsStore = defineStore("news", {
             }
         },
         async addLike(id_news, user_id) {
-            // const like = {
-            //     post_id: id_news,
-            //     likes_count: news.find(post => post.id === id_news).likes.length || 0,
-            // }
             const jwtToken = sessionStorage.getItem("jwtToken")
             if (!jwtToken) {
                 console.error("Отсутствует JWT-токен!")
@@ -168,15 +183,234 @@ export const useNewsStore = defineStore("news", {
                     }
                     console.log("добавлено")
                 }
-
+                this.getnews()
                 // const responseData = await response.json()
                 // console.log(responseData)
             } catch (error) {
                 console.error("Authentication error:", error)
             }
         },
-        // addNews(){
+        async addNews(newsTitle, newsText, picture, themes) {
+            const picture_url_obg = {
+                url: picture
+            }
+            const newsData = {
+                newsTitle: newsTitle,
+                newsText: newsText,
+                picture: picture_url_obg,
+                themes: themes.split(',').map(theme => ({ name: theme.trim() }))
+            }
+            console.log(newsData)
+            const jwtToken = sessionStorage.getItem("jwtToken")
+            if (!jwtToken) {
+                console.error("Отсутствует JWT-токен!")
+                return
+            }
+            try {
+                console.log(jwtToken)
+                const response = await fetch(
+                    `http://localhost:8085/news/create`,
+                    {
+                        method: "POST",
+                        headers: new Headers({
+                            Authorization: "Bearer " + jwtToken,
+                            "Content-Type": "application/json",
+                        }),
+                        body: JSON.stringify(newsData),
+                    }
+                )
+                if (!response.ok) {
+                    alert("Неправильный вход!")
+                    throw new Error("Authentication failed")
+                }
+                const responseData = await response.json()
+                console.log(responseData)
+                this.getnews()
+                this.closeModal()
+            } catch (error) {
+                console.error("Authentication error:", error)
+            }
+        },
+        async deleteNews() {
+            const picture_url_obg = {
+                url: picture
+            }
+            const newsData = {
+                newsTitle: newsTitle,
+                newsText: newsText,
+                picture: picture_url_obg,
+                themes: themes.split(',').map(theme => ({ name: theme.trim() }))
+            }
+            console.log(newsData)
+            const jwtToken = sessionStorage.getItem("jwtToken")
+            if (!jwtToken) {
+                console.error("Отсутствует JWT-токен!")
+                return
+            }
+            try {
+                console.log(jwtToken)
+                const response = await fetch(
+                    `http://localhost:8085/news/create`,
+                    {
+                        method: "POST",
+                        headers: new Headers({
+                            Authorization: "Bearer " + jwtToken,
+                            "Content-Type": "application/json",
+                        }),
+                        body: JSON.stringify(newsData),
+                    }
+                )
+                if (!response.ok) {
+                    alert("Неправильный вход!")
+                    throw new Error("Authentication failed")
+                }
+                const responseData = await response.json()
+                console.log(responseData)
+                this.getnews()
+                this.closeModal()
+            } catch (error) {
+                console.error("Authentication error:", error)
+            }
+        },
+        async getInfoNews(news_id) {
+            const jwtToken = sessionStorage.getItem("jwtToken")
+            if (!jwtToken) {
+                console.error("Отсутствует JWT-токен!")
+                return
+            }
+            try {
+                console.log(jwtToken)
+                const response = await fetch(
+                    `http://localhost:8085/news/${news_id}`,
+                    {
+                        method: "GET",
+                        headers: new Headers({
+                            Authorization: "Bearer " + jwtToken,
+                            "Content-Type": "application/json",
+                        }),
+                    }
+                )
+                if (!response.ok) {
+                    alert("Неправильный вход!")
+                    throw new Error("Authentication failed")
+                }
+                const responseData = await response.json()
+                this.news_for_edit = responseData
+                //sessionStorage.setItem('news_for_edit', JSON.stringify(responseData))
+                console.log('ssss')
+                console.log(this.news_for_edit)
+                // this.getnews()
+            } catch (error) {
+                console.error("Authentication error:", error)
+            }
+        },
+        async updateNews(newsTitle, newsText, picture, themes,  news_id) {
+            console.log(news_id)
+            const picture_url_obj = {
+                url: picture
+            }
+            const newsData = {
+                newsTitle: newsTitle,
+                newsText: newsText,
+                picture: picture_url_obj,
+                themes: themes.split(',').map(theme => ({ name: theme.trim() }))
+            }
+            console.log(newsData)
+            const jwtToken = sessionStorage.getItem("jwtToken")
+            if (!jwtToken) {
+                console.error("Отсутствует JWT-токен!")
+                return
+            }
+            try {
+                console.log(jwtToken)
+                const response = await fetch(
+                    `http://localhost:8085/news/${news_id}`,
+                    {
+                        method: "PUT",
+                        headers: new Headers({
+                            Authorization: "Bearer " + jwtToken,
+                            "Content-Type": "application/json",
+                        }),
+                        body: JSON.stringify(newsData),
+                    }
+                )
+                if (!response.ok) {
+                    alert("Неправильный вход!")
+                    throw new Error("Authentication failed")
+                }
+                this.getnews()
+                this.closeEdit()
+                // this.getnews()
+                //const responseData = await response.json()
+                //console.log(responseData)
+            } catch (error) {
+                console.error("Authentication error:", error)
+            }
+        },
+        showEdit(post_id){
+            this.edit_post_id = post_id
+            this.edit = true
+        },
+        closeEdit() {
+            this.edit = false
+        },
+        showModal(){
+            this.modal = true
+        },
+        closeModal(){
+            this.modal = false
+        },
+        async filterThemes(positive, negative) {
+            // {
+            //     "favoritesThemes" : [
+            //         {
+            //             "name": "Джаонозис"
+            //         }
+            //         ],
+            //     "forbiddenThemes" : [
+            //         {
+            //             "name": "Мустафар"
+            //         }
+            //         ]
+            // }
+            // const positive = positive.split(',').map(theme => ({ name: theme.trim() }))
+            // const negative = negative.split(',').map(theme => ({ name: theme.trim() }))
+            const themes = {
+                favoritesThemes: positive.split(',').map(theme => ({ name: theme.trim() })),
+                forbiddenThemes: negative.split(',').map(theme => ({ name: theme.trim() }))
+            }
+            console.log(themes)
+            const jwtToken = sessionStorage.getItem("jwtToken")
+            // if (!jwtToken) {
+            //     console.error("Отсутствует JWT-токен!")
+            //     return
+            // }
+            try {
+                console.log(jwtToken)
+                const response = await fetch(
+                    `http://localhost:8085/news/user-themes`,
+                    {
+                        method: "POST",
+                        headers: {
+                            'Content-Type': 'application/json'
+                            // Другие необходимые заголовки
+                          },
+                        body: JSON.stringify(themes),
+                    }
+                )
+                if (!response.ok) {
+                    alert("Неправильный вход!")
+                    throw new Error("Authentication failed")
+                }
+                const responseData = await response.json()
+                // this.filtr_news = responseData
+                console.log(responseData)
 
-        // }
+                this.getnews()
+            } catch (error) {
+                console.error("Authentication error:", error)
+            }
+        },
+
     },
 })
